requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==20.1.0
pytest==7.4.2
pytest-django==4.5.2
httpx==0.24.1
python-decouple